Fixed casting warning.

+++ mz_strm_mem.c
@@ -234,7 +234,7 @@ void mz_stream_mem_get_buffer_length(void *stream, int32_t *length)
     *length = mem->limit;
 }
 
-void mz_stream_mem_set_buffer_limit(void *stream, int64_t limit)
+void mz_stream_mem_set_buffer_limit(void *stream, int32_t limit)
 {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
     mem->limit = limit;
